+++ lib/screens/progression.dart
@@ -1,3 +1,4 @@
+import 'package:firebase_analytics/firebase_analytics.dart';
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:hugeicons/hugeicons.dart';
@@ -23,6 +24,10 @@ class _ProgressionState extends State<Progression> {
   @override
   void initState() {
     super.initState();
+    FirebaseAnalytics.instance.logScreenView(
+      screenName: '/progression',
+      screenClass: 'Progression',
+    );
     _colorListener = () {
       if (mounted) setState(() {});
     };